fix 写真撮影部分をcountdownに移植
