+++ app/(game)/[room_id]/countdown.tsx
@@ -1,4 +1,5 @@
 import { Button } from "@/components/Button/Button";
+import { PhotoTaker } from '@/components/Camera/PhotoTaker';
 import ScreenContainer from "@/components/ScreenContainer";
 import { Stack, useRouter } from "expo-router";
 import { Text } from "react-native";
@@ -13,6 +14,7 @@ export default function Countdown() {
       <Stack.Screen options={{ headerShown: false }} />
       <Text>カウント画面。多分ここでユーザーに写真撮ってもらうイメージになるのかな？</Text>
       <Button onPress={handlePress} text='次へ' />
+      <PhotoTaker themeHex="#bce2e8" />
     </ScreenContainer>
   )
 }
